requirements.txt
flask
torch
torchvision
pandas
numpy
pillow
